cargo/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ajax-datatable==4.4.4
django-crispy-forms==1.14.0
django-filter==22.1
gunicorn==20.1.0
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
